address comments.
